+++ src/views/settings/data.ts
@@ -1,0 +1,53 @@
+import { ClusterNode } from "./../../types/settings";
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+export const ClusterNodeRow = [
+  {
+    label: "Host",
+    value: "host",
+    width: 280,
+  },
+  {
+    label: "Port",
+    value: "port",
+    width: 180,
+  },
+  {
+    label: "Self",
+    value: "self",
+    width: 180,
+  },
+];
+
+export const SettingsTabs = [
+  {
+    label: "General",
+    value: "general",
+  },
+  {
+    label: "TTL",
+    value: "ttl",
+  },
+  {
+    label: "Cluster Nodes",
+    value: "clusterNodes",
+  },
+  {
+    label: "Dump Effective Configurations",
+    value: "dumpEffectiveConfigurations",
+  },
+];
